# Optimized box blur